+++ tests/testthat/test-geo_address_lookup.R
@@ -2,6 +2,7 @@ test_that("Returning empty query", {
   expect_message(geo_address_lookup("xbzbzbzoa aiaia", "R"))
 
   skip_if_api_server()
+  skip_on_cran()
 
   obj <- geo_address_lookup("xbzbzbzoa aiaia", "R")
 
@@ -20,6 +21,8 @@ test_that("Data format", {
 test_that("Checking query", {
   skip_if_api_server()
   skip_if_offline()
+  skip_on_cran()
+  
   expect_equal(ncol(geo_address_lookup(34633854, "W")), 4)
   expect_gt(ncol(geo_address_lookup(34633854, "W", full_results = TRUE)), 4)
   expect_equal(nrow(geo_address_lookup(34633854, "W",