+++ tests/testthat/test-geo_address_lookup_sf.R
@@ -1,6 +1,7 @@
 test_that("Returning empty query", {
   expect_message(geo_address_lookup_sf("xbzbzbzoa aiaia", "R"))
   skip_if_api_server()
+  skip_on_cran()
 
   obj <- geo_address_lookup_sf("xbzbzbzoa aiaia", "R")
 
@@ -12,6 +13,7 @@ test_that("Returning empty query", {
 test_that("Data format", {
   skip_if_api_server()
   skip_if_offline()
+  skip_on_cran()
   expect_true(is.data.frame(geo_address_lookup_sf(34633854, "W")))
   expect_s3_class(geo_address_lookup_sf(34633854, "W"), "sf")
 })
@@ -19,6 +21,7 @@ test_that("Data format", {
 test_that("Checking query", {
   skip_if_api_server()
   skip_if_offline()
+  skip_on_cran()
   expect_equal(ncol(geo_address_lookup_sf(34633854, "W")), 3)
   expect_gt(ncol(geo_address_lookup_sf(34633854, "W", full_results = TRUE)), 3)
   expect_equal(nrow(geo_address_lookup_sf(34633854, "W",