+++ tests/testthat/test-geo_amenity_sf.R
@@ -5,6 +5,7 @@ test_that("Returning empty query", {
   ))
 
   skip_if_api_server()
+  skip_on_cran()
 
   obj <- geo_amenity_sf(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
@@ -18,6 +19,7 @@ test_that("Returning empty query", {
 test_that("Data format", {
   skip_if_api_server()
   skip_if_offline()
+  skip_on_cran()
 
   expect_true(is.data.frame(geo_amenity_sf(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
@@ -33,6 +35,7 @@ test_that("Data format", {
 test_that("Checking query", {
   skip_if_api_server()
   skip_if_offline()
+  skip_on_cran()
 
   expect_equal(ncol(geo_amenity_sf(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),